+++ CloudHarvestApi/blueprints/tasks.py
@@ -1,9 +1,9 @@
 from CloudHarvestCoreTasks.blueprints import HarvestApiBlueprint
 from flask import Response, request
 from logging import getLogger
-from typing import Literal
 
-from CloudHarvestApi.blueprints.base import safe_jsonify
+from CloudHarvestCoreTasks.cache import CachedData
+from CloudHarvestApi.blueprints.base import safe_jsonify, safe_request_get_json, use_cache_if_valid
 from CloudHarvestApi.blueprints.home import not_implemented_error
 
 logger = getLogger('harvest')
@@ -12,6 +12,50 @@ tasks_blueprint = HarvestApiBlueprint(
     'tasks_bp', __name__,
     url_prefix='/tasks'
 )
+
+CACHED_TEMPLATES = CachedData(data=[], valid_age=0)
+
+
+@tasks_blueprint.route(rule='/await/<task_chain_id>', methods=['GET'])
+def await_task(task_chain_id: str) -> Response:
+    """
+    Awaits a task.
+
+    Arguments
+    task_chain_id: (str) The task chain ID (uuid4)
+
+    Returns
+    A response with the task chain results.
+    """
+
+    from datetime import datetime
+    from time import sleep
+
+    request_json = safe_request_get_json(request)
+
+    start_time = datetime.now()
+    timeout = request_json.get('timeout') or 120
+
+    while (datetime.now() - start_time).total_seconds() < timeout:
+        output = get_task_results(task_chain_id=task_chain_id).get_json()
+        reason = output.get('reason')
+
+        match reason:
+            case 'OK':
+                break
+
+            case _:
+                sleep(1)
+
+    else:
+        return safe_jsonify(
+            success=False,
+            reason='TIMEOUT',
+            result=None
+        )
+
+    return get_task_results(task_chain_id=task_chain_id)
+
 
 @tasks_blueprint.route(rule='/get_task_results/<task_chain_id>', methods=['GET'])
 def get_task_results(task_chain_id: str) -> Response:
@@ -55,39 +99,51 @@ def get_task_results(task_chain_id: str) -> Response:
             result=results
         )
 
-@tasks_blueprint.route(rule='/list_available_tasks/<task_type>', methods=['GET'])
-def list_available_tasks(task_type: Literal['reports', 'services']) -> Response:
+
+@tasks_blueprint.route(rule='/list_available_templates', methods=['GET'])
+@use_cache_if_valid(CACHED_TEMPLATES)
+def list_available_templates() -> Response:
     """
-    List the tasks available in the system.
-
-    Returns: A list of task models.
+    List the available task templates.
+    :return: A response.
     """
 
-    if task_type not in ('reports', 'services'):
-        return safe_jsonify(success=False, reason=f'Invalid task type: {task_type}', result=[])
+    from json import loads
+    from CloudHarvestCoreTasks.silos import get_silo
+    silo = get_silo('harvest-nodes')
 
-    from CloudHarvestCorePluginManager import Registry
+    client = silo.connect()
 
-    task_models = Registry.find(category=f'template_{task_type}', result_key='*', limit=None)
+    agents = client.keys('agent*')
 
-    result = {
-            'data': [
-                {
-                    'Name': model['name'],
-                    'Description': model['cls'][list(model['cls'].keys())[0]].get('description', 'Description not provided.'),
-                    'Tags': ', '.join(model.get('tags', []))
-                } for model in task_models
-            ],
-            'meta': {
-                'headers': ['Name', 'Tags', 'Description'],
-            },
-        }
+    reason = 'OK'
+    results = []
 
-    return safe_jsonify(
-        success=True,
-        reason='OK',
-        result=result
-    )
+    try:
+        for agent in agents:
+
+            agent_data = loads(client.get(name=agent))
+            agent_templates = agent_data.get('available_templates') or []
+
+            if isinstance(agent_templates, list):
+                results.extend(agent_templates)
+
+    except Exception as ex:
+        reason = f'Failed to list task results with error: {str(ex)}'
+        logger.error(reason)
+
+    finally:
+        results = sorted(list(set(results)))
+
+        # Update the CACHED_TEMPLATES so subsequent calls will be faster
+        CACHED_TEMPLATES.update(data=results, valid_age=300)
+
+        return safe_jsonify(
+            success=True,
+            reason=reason,
+            result=results
+        )
+
 
 @tasks_blueprint.route(rule='/list_task_results', methods=['GET'])
 def list_task_results() -> Response:
@@ -140,6 +196,7 @@ def list_task_queue() -> Response:
         result=results
     )
 
+
 @tasks_blueprint.route(rule='/escalate/<task_id>', methods=['GET'])
 def escalate_task(task_id: str) -> Response:
     """
@@ -152,7 +209,7 @@ def escalate_task(task_id: str) -> Response:
 
 
 @tasks_blueprint.route(rule='/queue/<priority>/<task_category>/<task_name>', methods=['POST'])
-def queue_task(priority: int, task_category: str, task_name: str) -> Response:
+def queue_task(priority: int, task_category: str, task_name: str, *args, **kwargs) -> Response:
     """
     Queues a task.
 
@@ -164,71 +221,76 @@ def queue_task(priority: int, task_category: str, task_name: str) -> Response:
 
     :return: A response.
     """
-    from CloudHarvestCorePluginManager import Registry
 
-    task_model = Registry.find(name=task_name, category=f'template_{task_category}', result_key='cls')
+    templates = list_available_templates().get_json().get('result')
 
-    if task_model:
-        task_model = task_model[0]
+    template_exists = False
+    for template in templates:
+        category, name = template.split('/')
+        category = category.replace('template_', '')
 
-        from CloudHarvestCoreTasks.silos import get_silo
-        silo = get_silo('harvest-task-queue')
-        client = silo.connect()
+        if category == task_category and name == task_name:
+            template_exists = True
+            break
 
-        from datetime import datetime, timezone
 
-        from uuid import uuid4
+    if not template_exists:
+        return safe_jsonify(
+            success=False,
+            reason=f'TEMPLATE NOT FOUND',
+            result=None
+        )
 
-        task = {
-            'id': str(uuid4()),
-            'priority': priority,
-            'name': task_name,
-            'category': task_category,
-            'model': task_model,
-            'config': dict(request.json),
-            'created': datetime.now(timezone.utc)
-        }
+    # The task is known to exist on some agent, therefore it can be queued
 
-        from json import dumps
-        payload = dumps(task, default=str)
+    from CloudHarvestCoreTasks.silos import get_silo
+    silo = get_silo('harvest-task-queue')
+    client = silo.connect()
 
-        # Create a unique name for the task
-        task_redis_name = f"task::{task['id']}"
+    from datetime import datetime, timezone
+    from uuid import uuid4
 
-        try:
+    task = {
+        'id': str(uuid4()),
+        'priority': priority,
+        'name': task_name,
+        'category': f'template_{task_category}',
+        'config': (dict(safe_request_get_json(request)) or {}) | kwargs,
+        'created': datetime.now(timezone.utc)
+    }
 
-            # Create the task queue item
-            client.setex(name=task_redis_name, value=payload, time=3600)
+    from json import dumps
+    payload = dumps(task, default=str)
 
-            # Now add the task to the queue
-            client.rpush(f"queue::{priority}", task_redis_name)
+    # Create a unique name for the task
+    task_redis_name = f"task::{task['id']}"
 
-        except Exception as ex:
-            reason = f'Failed to queue task {task_name} with error: {str(ex)}'
+    try:
+        # Create the task queue item
+        client.setex(name=task_redis_name, value=payload, time=3600)
 
-            # ROlLBACK
-            client.delete(name=task_redis_name)
-            client.lrem(name=f"queue::{priority}", value=task_redis_name)
+        # Now add the task to the queue
+        client.rpush(f"queue::{priority}", task_redis_name)
 
-        else:
-            reason = 'OK'
+    except Exception as ex:
+        reason = f'Failed to queue task {task_name} with error: {str(ex)}'
 
-        result = {
-            'success': reason == 'OK',
-            'reason': reason,
-            'result': {
-                'id': task['id'],
-                'priority': task['priority'],
-                'created': task['created'],
-            }
-        }
+        # ROlLBACK
+        client.delete(name=task_redis_name)
+        client.lrem(name=f"queue::{priority}", value=task_redis_name)
 
     else:
-        result = {
-            'success': False,
-            'reason': f'TaskChain {task_name} not found',
-            'result': {}
+        reason = 'OK'
+
+    result = {
+        'success': reason == 'OK',
+        'reason': reason,
+        'result': {
+            'id': task['id'],
+            'priority': task['priority'],
+            'created': task['created'],
         }
+    }
 
     return safe_jsonify(
         success=result['success'],